requirements.txt
asgiref==3.7.2
attrs==23.1.0
backports.zoneinfo==0.2.1
black==23.9.1
click==8.1.7
colorama==0.4.6
coverage==7.3.2
Django==4.2.6
django-js-asset==2.1.0
django-mptt==0.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
exceptiongroup==1.1.3
factory-boy==3.3.0
Faker==19.10.0
importlib-resources==6.1.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
pluggy==1.3.0
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
zipp==3.17.0
